pyproject.toml
[project]
name = "pythonproject1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "selenium>=4.40.0",
]
